Organize components from `EMExpandedLayoutView` into their own files.
